refactor: decouple feedback repository
