+++ src/libs/modules/database/repositories/feedback/feedback.interface.ts
@@ -1,0 +1,7 @@
+import { Feedback, Prisma } from '@prisma/client';
+
+export interface FeedbackRepository {
+  create(input: Prisma.FeedbackCreateArgs): Promise<Feedback>;
+}
+
+export const FeedbackRepository = Symbol('FeedbackRepository');
